[23.10.07]Signup Component updated
